chore: remove unnecessary build
