+++ configs/test-utils/package.json
@@ -13,34 +13,12 @@
   },
   "sideEffects": false,
   "type": "module",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    },
-    "./package.json": "./package.json"
-  },
-  "main": "dist/index.cjs",
-  "types": "dist/index.d.ts",
+  "main": "src/index.ts",
   "scripts": {
-    "build": "tsup",
-    "build:watch": "tsup --watch",
-    "clean": "rimraf ./dist",
     "lint": "eslint \"**/*.ts*\"",
-    "lint:attw": "attw --pack",
-    "lint:pub": "publint --strict",
-    "prepack": "pnpm build",
     "type:check": "tsc --noEmit"
   },
   "devDependencies": {
-    "@suspensive/tsup": "workspace:*",
-    "@types/node": "^18.16.2",
     "@types/react": "^18.2.0",
     "react": "^18.2.0"
   },
