+++ configs/test-utils/src/ThrowError.tsx
@@ -1,4 +1,4 @@
-import type { PropsWithChildren, ReactNode } from 'react'
+import type { PropsWithChildren } from 'react'
 import { useEffect, useState } from 'react'
 
 const useSetTimeout = (fn: (...args: []) => void, delay: number) =>
@@ -6,23 +6,6 @@ const useSetTimeout = (fn: (...args: []) => void, delay: number) =>
     const timeout = setTimeout(fn, delay)
     return () => clearTimeout(timeout)
   }, [fn, delay])
-
-const suspendIsNeed = { current: true }
-type SuspendProps = { during: number; toShow?: ReactNode }
-export const Suspend = ({ during, toShow }: SuspendProps) => {
-  if (suspendIsNeed.current) {
-    throw new Promise((resolve) =>
-      setTimeout(() => {
-        suspendIsNeed.current = false
-        resolve('resolved')
-      }, during)
-    )
-  }
-  return <>{toShow}</>
-}
-Suspend.reset = () => {
-  suspendIsNeed.current = true
-}
 
 const throwErrorIsNeed = { current: false }
 type ThrowErrorProps = PropsWithChildren<{ message: string; after?: number }>
@@ -48,10 +31,3 @@ export const ThrowNull = ({ after, children }: ThrowNullProps) => {
 ThrowError.reset = () => {
   throwErrorIsNeed.current = false
 }
-
-export const delay = (ms: number) => new Promise((resolve) => setTimeout(() => resolve('done'), ms))
-
-export const TEXT = 'TEXT' as const
-export const ERROR_MESSAGE = 'ERROR_MESSAGE' as const
-export const FALLBACK = 'FALLBACK' as const
-export const MS_100 = 100 as const
