+++ configs/test-utils/src/index.ts
@@ -1,0 +1,7 @@
+export { Suspend } from './Suspend'
+export { ThrowError, ThrowNull } from './ThrowError'
+export { delay } from './delay'
+export const TEXT = 'TEXT' as const
+export const ERROR_MESSAGE = 'ERROR_MESSAGE' as const
+export const FALLBACK = 'FALLBACK' as const
+export const MS_100 = 100 as const
