--- configs/test-utils/tsup.config.ts
@@ -1,4 +1,0 @@
-import { options } from '@suspensive/tsup'
-import { defineConfig } from 'tsup'
-
-export default defineConfig(options)
